useFetchSnsPosts hook 만들고 적용
